Create logout confirmation dialog
#6
