+++ app/src/main/java/by/alexandr7035/affinidi_id/data/AuthRepository.kt
@@ -1,5 +1,6 @@
 package by.alexandr7035.affinidi_id.data
 
+import by.alexandr7035.affinidi_id.data.model.log_out.LogOutModel
 import by.alexandr7035.affinidi_id.data.model.sign_in.SignInModel
 import by.alexandr7035.affinidi_id.data.model.sign_up.SignUpConfirmationModel
 import by.alexandr7035.affinidi_id.data.model.sign_up.SignUpModel
@@ -23,4 +24,6 @@ interface AuthRepository {
     fun checkIfAuthorized(): Boolean
 
     fun saveUserName(userName: String)
+
+    suspend fun logOut(): LogOutModel
 }
