+++ app/src/main/java/by/alexandr7035/affinidi_id/data/implementation/AuthRepositoryImpl.kt
@@ -6,6 +6,7 @@ import by.alexandr7035.affinidi_id.core.extensions.debug
 import by.alexandr7035.affinidi_id.data.ApiService
 import by.alexandr7035.affinidi_id.data.AuthDataStorage
 import by.alexandr7035.affinidi_id.data.AuthRepository
+import by.alexandr7035.affinidi_id.data.model.log_out.LogOutModel
 import by.alexandr7035.affinidi_id.data.model.sign_in.SignInModel
 import by.alexandr7035.affinidi_id.data.model.sign_in.SignInRequest
 import by.alexandr7035.affinidi_id.data.model.sign_in.SignInResponse
@@ -148,5 +149,48 @@ class AuthRepositoryImpl @Inject constructor(
     private fun saveAuthData(userDid: String, accessToken: String) {
         authDataStorage.saveDid(userDid)
         authDataStorage.saveAccessToken(accessToken)
+    }
+
+    override suspend fun logOut(): LogOutModel {
+        try {
+            val res = apiService.logOut(authDataStorage.getAccessToken() ?: "")
+
+            if (res.isSuccessful) {
+                Timber.debug("LOGOUT SUCCESSFUL")
+
+                authDataStorage.saveUserName(null)
+                authDataStorage.saveAccessToken(null)
+                authDataStorage.saveDid(null)
+
+                return LogOutModel.Success()
+            }
+            else {
+                return when (res.code()) {
+                    401 -> {
+                        // Authorization token already unactual
+                        // Just clear it and return success logout
+                        authDataStorage.saveUserName(null)
+                        authDataStorage.saveAccessToken(null)
+                        authDataStorage.saveDid(null)
+                        return LogOutModel.Success()
+                    }
+                    else -> {
+                        LogOutModel.Fail(ErrorType.UNKNOWN_ERROR)
+                    }
+                }
+            }
+        }
+        // Handled in ErrorInterceptor
+        catch (appError: AppError) {
+            appError.printStackTrace()
+            Timber.debug("ERRORTYPE ${appError.errorType.name}")
+            return LogOutModel.Fail(appError.errorType)
+        }
+        // Unknown exception
+        catch (e: Exception) {
+            Timber.debug("LOGOUT FAILED other exception $e")
+            e.printStackTrace()
+            return LogOutModel.Fail(ErrorType.UNKNOWN_ERROR)
+        }
     }
 }
