+++ app/src/main/java/by/alexandr7035/affinidi_id/presentation/logout/LogoutFragment.kt
@@ -1,0 +1,61 @@
+package by.alexandr7035.affinidi_id.presentation.logout
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.viewModels
+import androidx.navigation.fragment.findNavController
+import by.alexandr7035.affinidi_id.R
+import by.alexandr7035.affinidi_id.core.ErrorType
+import by.alexandr7035.affinidi_id.core.extensions.navigateSafe
+import by.alexandr7035.affinidi_id.core.extensions.showToast
+import by.alexandr7035.affinidi_id.data.model.log_out.LogOutModel
+import by.alexandr7035.affinidi_id.databinding.FragmentLoginBinding
+import by.alexandr7035.affinidi_id.databinding.FragmentLogoutBinding
+import by.alexandr7035.affinidi_id.presentation.profile.ProfileFragmentDirections
+import by.kirich1409.viewbindingdelegate.viewBinding
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class LogoutFragment : BottomSheetDialogFragment() {
+
+    private val binding by viewBinding(FragmentLogoutBinding::bind)
+    private val viewModel by viewModels<LogoutViewModel>()
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_logout, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        viewModel.logOutLiveData.observe(viewLifecycleOwner, { logout ->
+            when (logout) {
+                is LogOutModel.Success -> {
+                    requireContext().showToast(getString(R.string.successful_logout))
+                    findNavController().navigateSafe(LogoutFragmentDirections.actionLogoutFragmentToLoginFragment())
+                }
+                is LogOutModel.Fail -> {
+                    when (logout.errorType) {
+                        ErrorType.FAILED_CONNECTION -> {
+                            requireContext().showToast(getString(R.string.error_failed_connection))
+                        }
+                        ErrorType.UNKNOWN_ERROR -> {
+                            requireContext().showToast(getString(R.string.error_unknown))
+                        }
+                        else -> {
+                            requireContext().showToast(getString(R.string.error_unknown))
+                        }
+                    }
+                }
+            }
+        })
+
+        binding.logoutBtn.setOnClickListener {
+            viewModel.logOut()
+        }
+    }
+}
