+++ app/src/main/res/drawable/background_bottom_dialog.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <solid android:color="@color/white"/>
+    <corners android:topLeftRadius="@dimen/bottom_dialog_corner_radius" android:topRightRadius="@dimen/bottom_dialog_corner_radius"/>
+</shape>
