+++ app/src/main/res/layout/fragment_logout.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    style="@style/BottomDialogBackground"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:orientation="vertical"
+    tools:context=".presentation.logout.LogoutFragment">
+
+    <TextView
+        style="@style/CardHeaderText"
+        android:textAlignment="center"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:text="@string/want_to_log_out" />
+
+    <TextView
+        style="@style/TextViewExplanation"
+        android:layout_marginTop="8dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/log_out_explanation"/>
+
+    <Button
+        android:id="@+id/logoutBtn"
+        style="@style/PrimaryButton"
+        android:layout_marginTop="8dp"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/log_out"/>
+
+</LinearLayout>
